Create R04_STR04_J.java
Non-compliant Code Example

This non-compliant code example reads a byte array and converts it into a String using the platform's default character encoding. If the byte array does not represent a string, or if it represents a string that was encoded using other than the default encoding, the resulting String is likely to be incorrect. The behavior resulting from malformed-input and unmappable-character errors is unspecified.